+++ releases/ironwood/2/oee/CHANGELOG.md
@@ -9,6 +9,10 @@ release.
 
 ## [Unreleased]
 
+### Changed
+
+- Update default password policy
+
 ## [ironwood.2-oee-1.0.0] - 2020-09-10
 
 - First release of an `ironwood.2-oee` Docker image.